Camera HAL3: cast long int to int for liblog to fix compiling

+++ camera/QCamera2/HAL/QCameraPostProc.cpp
@@ -2365,11 +2365,11 @@ void *QCameraPostProcessor::dataSaveRoutine(void *data)
                         if ((ssize_t)job_data->out_data.buf_filled_len != written_len) {
                             ALOGE("%s: Failed save complete data %d bytes "
                                   "written instead of %d bytes!",
-                                  __func__, written_len,
+                                  __func__, (int) written_len,
                                   job_data->out_data.buf_filled_len);
                         } else {
                             CDBG_HIGH("%s: written number of bytes %d\n",
-                                __func__, written_len);
+                                __func__, (int) written_len);
                         }
 
                         close(file_fd);
